create module convs inside layers and simplify code with modularity and using forward pass of convolution layer and convtransposed layer. ConvTransposed TODO.
